fix: XPath div operator returns decimal for integer operands
XPath 3.1 div must never return integer; integer operands are promoted
to decimal before division. This fixes fold-left-019 and any other
expression relying on decimal division semantics.

- VmEngine.Divide: removed integer/integer branch, always promotes to decimal
- XPathOptimizer: integer div constant fold now produces DecimalLiteralNode
- Updated unit tests to expect DecimalValue instead of IntegerValue

+++ tests/Bosak.XPath.Runtime.Tests/VmEngineTests.cs
@@ -163,7 +163,7 @@ public class VmEngineTests
     public void Eval_Division()
     {
         var result = Evaluate("20 div 4");
-        Assert.Equal(5, result.IntegerValue);
+        Assert.Equal(5.0m, result.DecimalValue);
     }
 
     [Fact]
